2.16 tehty
Tietojen poistaminen valmis

+++ puhelinluettelo/src/App.js
@@ -54,15 +54,21 @@ class App extends React.Component {
     }
 
 
-  removePerson = (id) =>{
+  removePerson = (id, person) =>{
     return () => {
       console.log(`should ${id} be removed?`)
-      personService
-      .deletePerson(id)
-      .then(response => {
-        console.log('promise fulfilled')
-        this.setState({ persons : response})
-      })
+      if (window.confirm(`Haluatko varmasti poistaa ${person}`)){
+        personService
+        .deletePerson(id)
+        .then( () => {
+          personService
+          .getAll()
+          .then(response => {
+            console.log('promise fulfilled')
+            this.setState({ persons : response})
+          })
+        })
+      }
     }
   }
 
@@ -98,7 +104,7 @@ class App extends React.Component {
         <RajaaNaytettavia filter= {this.state.filter} handlefilterChange= {this.handlefilterChange}/>
         <h2>Numerot</h2>
         <div>
-          {found.map(item => <YksittaisenHenkilonTiedot key={item.name} name={item.name} number={item.number} removePerson = {this.removePerson(item.id)}/>)}
+          {found.map(item => <YksittaisenHenkilonTiedot key={item.name} name={item.name} number={item.number} removePerson = {this.removePerson(item.id, item.name)}/>)}
         </div>
 
       </div>
